common core

+++ easy-common-core/src/main/java/com/yffd/easy/common/core/page/PageParam.java
@@ -19,24 +19,24 @@ public class PageParam implements Serializable {
 	private static final long serialVersionUID = 6844568491810672063L;
 
 	/** 开始页码为1*/
-    public static final Long PAGE_NUM_START = 1L;
+    public static final Integer PAGE_NUM_START = 1;
     /** 每页最小记录数(15) */
-    public static final Long PAGE_LIMIT_MIN = 1L;
+    public static final Integer PAGE_LIMIT_MIN = 1;
     /** 每页最大记录数(100) */
-    public static final Long PAGE_LIMIT_MAX = 100L;
+    public static final Integer PAGE_LIMIT_MAX = 100;
     
-    private Long pageNum; 		// 当前页数
-    private Long pageLimit; 	// 每页记录数
-    private Long totalRecord; 	// 总记录数
+    private Integer pageNum; 		// 当前页数
+    private Integer pageLimit; 	// 每页记录数
+    private Integer totalRecord; 	// 总记录数
     
-//    private Long startIndex; 	// 每页开始行号
-//    private Long endIndex; 	// 每页结束行号
-//    private Long totalPage; 	// 总页数
+//    private Integer startIndex; 	// 每页开始行号
+//    private Integer endIndex; 	// 每页结束行号
+//    private Integer totalPage; 	// 总页数
     
     /** 是否进行自动count计算总条数，默认为自动计算：true:是，false:否 */
 //    private boolean autoCount = true;
     
-	public PageParam(Long pageNum, Long pageLimit) {
+	public PageParam(Integer pageNum, Integer pageLimit) {
 		//当前页码不能小于1，如果小于将返回最小页码
 		pageNum = pageNum < 1 ? PAGE_NUM_START : pageNum;
 		// pageLimit in [ PAGE_LIMIT_MIN, PAGE_LIMIT_MAX ] 范围
@@ -51,7 +51,7 @@ public class PageParam implements Serializable {
 	 * @author  zhangST
 	 * @return
 	 */
-	public Long getPageNum() {
+	public Integer getPageNum() {
 		return this.pageNum;
 	}
 
@@ -61,7 +61,7 @@ public class PageParam implements Serializable {
 	 * @author  zhangST
 	 * @return
 	 */
-	public Long getPageLimit() {
+	public Integer getPageLimit() {
 		return this.pageLimit;
 	}
 
@@ -71,7 +71,7 @@ public class PageParam implements Serializable {
 	 * @author  zhangST
 	 * @return
 	 */
-	public Long getTotalRecord() {
+	public Integer getTotalRecord() {
 		return totalRecord;
 	}
 
@@ -81,7 +81,7 @@ public class PageParam implements Serializable {
 	 * @author  zhangST
 	 * @param totalRecord
 	 */
-	public void setTotalRecord(Long totalRecord) {
+	public void setTotalRecord(Integer totalRecord) {
 		this.totalRecord = totalRecord < 0 ? 0 : totalRecord;
 	}
 
@@ -91,7 +91,7 @@ public class PageParam implements Serializable {
 	 * @author  zhangST
 	 * @return
 	 */
-	public Long getTotalPage() {
+	public Integer getTotalPage() {
 		return this.getTotalRecord() % this.getPageLimit() == 0 ? 
 				this.getTotalRecord() / this.getPageLimit() : this.getTotalRecord() / this.getPageLimit() + 1;
 	}
@@ -102,7 +102,7 @@ public class PageParam implements Serializable {
 	 * @author  zhangST
 	 * @return
 	 */
-	public Long getStartIndex() {
+	public Integer getStartIndex() {
 		return (this.getPageNum() - 1) * this.getPageLimit();
 	}
 
@@ -112,7 +112,7 @@ public class PageParam implements Serializable {
 	 * @author  zhangST
 	 * @return
 	 */
-	public Long getEndIndex() {
+	public Integer getEndIndex() {
 		return this.getPageNum() * this.getPageLimit();
 	}
